updates tex.file
difference depending on ICOS / non-ICOS stations.

table was discussed for the settings. included necessary ones (date-range, hour(s), map degrees and distances

previously hard coded ex 2018 and that the seasonal variations table is december 2017 to december 2018. Now 2018 removed and the caption for the seasonal table changes depending on setting for specific run. consider if need a specific tex-file layout for the 2018 runs for the landing pages.

could still improve on the tex-code... such as replace the multiple "bigskip".

note: uses specifications.tex (same for all output files) and the ICOS logo which needs to be uploaded (to where the .nc-files and such are uploaded?)