fixes issue #73
In the settings used to produce error #73 , the station actually has some footprints for year 2006 (despite key-error). No months are listed for 2006. Could look into why the created stilt dictionary (found in the object for each station characterization run under .settings['stilt']) has it as an available year but not month.