updated population data
Previously always used population data from 2011. Now there was a 2018 release and the .nc file is updated so it now has two variables rather than only one: 2011 and 2018. Which data is used depends on the start year of the station characterization. If the start year is less than 2015, 2011 will be used, and otherwise the data for 2018. Could consider a linear interpolation to determine values for the gap years.
In order to work, GEOSTAT_population_2011.nc needs to be replaced with GEOSTAT_population_2011_2018.nc in /data/project/stc/