updates from project folder
network_characterization_functions include functions that are only relevant for notebooks in the projects folder (LULC change and data-time select). Consider removing these in the future.
Another change is that data is loaded from common folder /data/project/stc/ --> many of the datasets used are also used in the stc project. There are still some files that are uploaded which the users can se, consider creating a specific folder for these (data/project/networkc?)