optimize network first launch!
Tool that uses a lot of code from other notebook tools:
Bokeh multiple variables graph with variables including land cover, but also values from the original station characterization seasonal table such as anthropogenic contribution and population sensitivity.
The variables can be selected as well as the stations (this is new). Also, prioritizing the variables to generate a combined score (weighted differently - same principles as suitability analysis). This is in progress.