slight update date range
Previously specifying (for example) start date 2018-01-01 to 2018-12-31 would aggregate 2913 as opposed to 2920 footprints. Last date stamp was 2018-12-31 00:00 rather than 2018-12-31 21:00. This code includes all the footprints of the final day. In seasonal table and precomputed footprints (used in network characterization) the date range for the full year is was already specified in this way.