no longer 10 m inlet height JFJ
It will be 5 meters from now on (and the 5-meter inlet height files have the older measurements). See message from Susanne (radiocarbon lab):
The height of the 14C sampling at JFJ was adapted from 10 to 5m in October 2021, following the reworking of Ingeborg on the longterm JFJ dataset, which was published at this time.
To be consistent we had to change the sampling height also for the ICOS release.

So please take into account the 5m sampling height.