resolved: setting upload not working in JupyterLab
when "upload" is used in the station characterization tool in the current image, "uploaded_file" is read as a dictionary in gui.py. In the new image, it is rather read as a tuple (code still allow for the "old reading" of the settings file --> in case read as a dictionary). Furthermore, when the file content is accessed in the new image, it is in a "memoryview object", but can be used after converting it with tobytes()