"improved monitoring potential"
Monitoring potential is calculated from taking the "even footprint" minus the footprint. To be able to compare how much the monitoring potential is improving compared to the "current network", the spread out "current network" is used also for the monitoring potential of the extended network (last cell in notebook).

For Lithuania and other countries with barely any sensing in the "current network" the even footprint becomes VERY small compared to the extended network which it is now compared to (even footprint minus footprint).

Only cells where the even footprint have higher value than the (extended) footprint will show up in the monitoring potential map. The condition for the message ("Sensing of X in country Y is very low and it is not feasible to show monitoring potential of") is not appropriate here and I suggest this message: "The improved monitoring of X in country Y is almost infinite in the extended network compared to the current network. To see the monitoring potential of the extended network, select it when prompted for "Network name" in section 3.2. and run the tool without extension."

Also some more explaining when that part of the tool is introduced:
Figure 6: Monitoring potential maps based on the extended network
The selected network is used as a reference (used to create the "equal views") which means that the monitoring potential maps for the current and extended networks have the same colorbars and can be used to see how the monitoring potential is improved.

Due to our definition of monitoring potential, in case of a country where the sensing of the selected network is close to zero the improvement in monitoring is almost infinite. No monitoring potential maps will be shown in these cases and users will rather get the message that to "To see the monitoring potential of the extended network, select it when prompted for "Network name" in section 3.2. and run the tool without extension."