Ute feedback: signals table "other" category
Not including the fractional land cover masks for ocean and unknown which leads to a discrepancy compared to the "total GEE/respiration",

Ocean is not associated with any fluxes in VPRM. The values associated with ocean comes from grid cells that have both ocean and land in them. There the fractional ocean mask “gets some fluxes” assigned.

For consistency:
- "other" category include also "ocean" and "unknown". put last in the table, closest to the "total".