update std bar threshold 2000-2020 referece period
In case 16/21 values for a certain date during the reference period are flagged, the standard deviation bar is not included. This only applies when the users look at the "monthly scope" (and see anomalies in for instance January a specific year, as opposed to the different months during that year - "yearly scope".
If the reference period is 2010-2020 it is rather 6/11 values that need to be flagged.
The different thresholds can be discussed, but they should be different for the different reference periods.