calculation of averages
Update to calculation of averages:

There was a place where values in the timeseries were disregarded because they were too few - less than 7 - in a month (and should not be displayed as a monthly average). However, they should still be used to calculate the yearly average.

What complicates it is the fact that I need three dataframes with the final data:
- one where the averages are calculated
- one where there are nans in place where months have too few values (mentioned about)
- one (for display) where all nans are set to zero instead.

Now all this happens in the plot_interface_anomaly py-file rather than split between this file and the "gui_anomaly file.