90% of footprints required for seasonal table
Previously, we only checked whether there were footprints for all 12 months. In a handful of cases, users have computed just a couple of footprints in each month. Despite this, an annual value was calculated and displayed in the seasonal table. Now, we ensure that there are at least 90% of the footprints before the averages are calculated (in cases where these values are computed, as opposed to loading them from a file)