seasonal variation table values: bug fixed
The anthropogenic component, and its variation over the different seasons:  "residential" and "other_categories" data from winter was wrongfully used for all seasons. Additionally, winter averages for these two categories were alone was used to calculate the yearly average anthropogenic component. This error was introduced after splitting the STILT categories from "others" to "residential" and "other_categories", and was not present in the pre-computed values.