multiple_variables graph update + refactor
The code for generating the multiple variables graph has been refactored.

There is one update that influences the result: pre-computed annual averages for the start year of the selected date range were used regardless of the date range. I only verified that the user had chosen all hours of the day. This will not make a difference for the station characterization documents, as they all used full years. In this updated version, the pre-computed values are only used in case of a user-defined date range that result in 2920 or 2928 (leap year) timestamps.