ex2_station: popup, legend and markercluster
It's a hard to find the balance between providing a simple example and a rich map... Here is my suggestion.

Perhaps we could consider adding something like this to the library? It would be great to input a list of stations (preferably just IDs) and generate a similar map. If we do this, we might want to center the map based on the latitudes and longitudes of the input stations and also adjust the zoom based on the bounding coordinates.