Countries shown in more tables
Country column now included in all dataframe column. For the similarity matrix it is not included as "city_name" is used as index and other than that it is a matrix with the cities. Could consider combining the city_name and Country column for the index column, but this may affect later steps.