Note about restarting kernel
Note about restarting the kernel. Inconsistent issue:
"When I change my mind after having run through the whole city characteristics notebook and change the selection of cities in “create a subset of cities” and then continue running the notebook (again), everything works fine until I upload the weights. In the corresponding output table the list of cities is not updated. This is not a major issue because one could just restart the notebook kernel and run again. All other save buttons give feedback."